Added Profile Component
