+++ src/webapp/myreact/src/components/ProfileComponent.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+
+class Profile extends Component {
+
+    constructor(props) {
+        super(props);
+    }
+
+    render() {
+        return (
+            <div>Test Profile</div>
+        );
+    }
+}
+
+export default Profile;
